feat(asset-import): Move asset stream function to assets module
