+++ lib/tasks/push-to-space/push-to-space.js
@@ -1,7 +1,3 @@
-import fs from 'fs'
-import { promisify } from 'util'
-import { join } from 'path'
-
 import Listr from 'listr'
 import verboseRenderer from 'listr-verbose-renderer'
 
@@ -11,8 +7,6 @@ import { wrapTask } from 'contentful-batch-libs/dist/listr'
 import * as assets from './assets'
 import * as creation from './creation'
 import * as publishing from './publishing'
-
-const stat = promisify(fs.stat)
 
 const DEFAULT_CONTENT_STRUCTURE = {
   entries: [],
@@ -74,19 +68,6 @@ export default function pushToSpace ({
 
   listrOptions = listrOptions || {
     renderer: verboseRenderer
-  }
-
-  const getAssetStreamForURL = async (url) => {
-    const [, assetPath] = url.split('//')
-    const filePath = join(assetsDirectory, assetPath)
-    try {
-      await stat(filePath)
-      return fs.createReadStream(filePath)
-    } catch (err) {
-      const error = new Error('Cannot open asset from filesystem')
-      error.filePath = filePath
-      throw error
-    }
   }
 
   const destinationDataById = {}
@@ -208,7 +189,7 @@ export default function pushToSpace ({
             allPendingUploads.push(requestQueue.add(async () => {
               try {
                 logEmitter.emit('info', `Uploading Asset file ${file.upload}`)
-                const assetStream = await getAssetStreamForURL(file.upload)
+                const assetStream = await assets.getAssetStreamForURL(file.upload, assetsDirectory)
                 const upload = await ctx.environment.createUpload({
                   fileName: asset.transformed.sys.id,
                   file: assetStream
